--- src/main/java/ar/edu/unlam/tallerweb1/servicios/ServicioObtenerListaPerdida.java
@@ -1,10 +1,0 @@
-package ar.edu.unlam.tallerweb1.servicios;
-
-import java.util.List;
-
-import ar.edu.unlam.tallerweb1.modelo.PostMascotaPerdida;
-
-public interface ServicioObtenerListaPerdida {
-	
-	List<PostMascotaPerdida> obtenerListaPerdida();
-}
