+++ src/main/java/ar/edu/unlam/tallerweb1/servicios/ServicioPost.java
@@ -1,7 +1,10 @@
 package ar.edu.unlam.tallerweb1.servicios;
 
+import java.util.List;
+
 import ar.edu.unlam.tallerweb1.modelo.PostMascotaPerdida;
 
-public interface ServicioGuardarPost  {
+public interface ServicioPost  {
 	void guardarPost(PostMascotaPerdida post);
+	List<PostMascotaPerdida> obtenerListaPerdida();
 }
